backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# RAG & AI
llama-index==0.10.0
llama-index-vector-stores-chroma==0.1.4
llama-index-llms-anthropic==0.1.5
llama-index-embeddings-huggingface==0.1.4
anthropic==0.18.0

# Vector Database
chromadb==0.4.22
sentence-transformers==2.3.1

# HTTP & API
httpx==0.26.0
requests==2.31.0
aiohttp==3.9.1

# Data Processing
numpy==1.24.3
pandas==2.1.4

# Environment
python-dotenv==1.0.0

# Google Cloud
google-cloud-storage==2.14.0
google-cloud-run==0.9.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
tenacity==8.2.3
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Logging
loguru==0.7.2

# Security
cryptography==42.0.0
python-jose[cryptography]==3.3.0
